+++ R/app_ui.R
@@ -47,7 +47,12 @@ TutorialUI <- function() {
 AttributionResultsUI <- function() {
   fluidPage(
     br(),
-    fixedRow(splitLayout(cellWidths = c("25%", "17.5%", "20%"),
+    fluidRow(actionButton(inputId = "analyzeMoreSample", 
+                          label = "Analyze another sample",
+                          style = "color: #fff; background-color: #337ab7;
+                             border-color: #2e6da4;")),
+    br(),
+    fluidRow(splitLayout(cellWidths = c("25%", "17.5%", "20%"),
                          downloadButton(outputId = "downloadExposureTable", 
                                         label = "Download exposure counts"),
                          downloadButton(outputId = "downloadPdf",