+++ NEWS.md
@@ -3,6 +3,8 @@
 * Added progress message if using bootstrapping to remove some of the signatures in 
 the beginning.
 
+* Added another actionButton "Analyze another sample" on the attribution results page.
+
 ## Changed
 * Do not create "showSpectraTab", "sigAttributionTab" and
 "attributionResultsTab" on initialization. Use `inserTab` instead to show the